[ALPS09708637] gz: fix overflowed constant
[Description]
fix overflowed constant

[Test]
build pass, boot pass

MTK-Commit-Id: 1a628bb7748d2e3f0e58c9686a7c47a11876442a

CR-Id: ALPS09708637
Feature: GenieZone

+++ drivers/misc/mediatek/geniezone/gz_main.c
@@ -985,7 +985,7 @@ static long _gz_ioctl(struct file *filep, unsigned int cmd, void __user *arg,
 		shm_data.shm_handle = shm_handle;
 
 		/* copy result back to user */
-		shm_data.session = ret;
+		shm_data.session = (uint32_t)ret;
 		err = copy_to_user(arg, &shm_data, sizeof(shm_data));
 		if (err) {
 			KREE_ERR("[%s]copy_to_user fail(0x%x)\n", __func__,
